Fix wordList example
The `wordList` example was quite broken because it incorrectly counted the
size of the `wordList` array and then didn't put a terminating `NUL`
character at the end of the `letterlist` variable.  Fixing it without
adding more C++-isms was difficult, so I added a `vector` of `string`s
and then fixed up some compiler issues.

This may not be the proper fix if we wish to introduce `vector` after
this chapter is completed.  If so, then a different example should
probably be used.